tests/requirements-test.txt
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-cov>=4.1.0
httpx>=0.24.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
pydantic>=2.1.1
uvicorn>=0.23.1
fastapi>=0.100.0
